requirements.txt
arrow==0.17.0
beautifulsoup4==4.9.3
binaryornot==0.4.4
briefcase==0.3.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
gitdb==4.0.5
GitPython==3.1.11
idna==2.10
importlib-metadata==2.0.0
Jinja2==2.11.2
jinja2-time==0.2.0
MarkupSafe==1.1.1
poyo==0.5.0
pycparser==2.20
python-dateutil==2.8.1
python-slugify==4.0.1
pythonnet==2.5.1
requests==2.24.0
six==1.15.0
smmap==3.0.4
soupsieve==2.0.1
text-unidecode==1.3
toga-core==0.3.0.dev25
toga-winforms==0.3.0.dev25
toml==0.10.1
travertino==0.1.3
urllib3==1.25.11
zipp==3.4.0
